backend/requirements.txt
alembic==1.7.7
apscheduler==3.8.1
atlassian-python-api==3.22.0
feedparser==6.0.8
flask==2.1.0
flask-cors==3.0.10
flask-jwt-extended==4.4.1
flask-limiter==2.4.6
flask-login==0.5.0
pdpyras==4.5.0
psycopg2==2.9.2
pyjokes==0.6.0
python-dotenv==0.19.2
requests==2.26.0
slack_bolt==1.13.0
slack_sdk==3.15.2
sqlalchemy==1.4.29
waitress==2.1.1
werkzeug~=2.0.0
